Refactor code structure for improved readability and maintainability
- Remove `public` folder, not needed at this point in time
- Move ban list and weird line list outside to not muddy the text cleaner scripts
- Stop using langdetect, using ban list to detect non-English words is faster
- Detect if line is pure number or a fraction like "5/2"

+++ text_cleaners/remakeTestFile.py
@@ -4,27 +4,25 @@ from changeFileExtension import changeFileExtension
 from addHeading import addHeading
 from deleteLine import deleteLine
 from deleteCodeBlock import deleteCodeBlock
-from generateAudio import TTSProcessor, generateTTS
 
 import os
 from tqdm import tqdm
 
-savePath = os.path.join(os.path.dirname(os.path.dirname(__file__)), "test")
-weirdPath = os.path.join(os.path.dirname(__file__), "weirdList.txt")
-fullPath = os.path.join(os.path.dirname(os.path.dirname(__file__)), "full.txt")
+rootProjectPath = os.path.dirname(os.path.dirname(__file__))
+
+savePath = os.path.join(rootProjectPath, "test")
+weirdPath = os.path.join(rootProjectPath, "weird_line.txt")
+fullPath = os.path.join(rootProjectPath, "full.txt")
 
 
 def remakeTestFile():
     deleteTestFiles(savePath)
     print("Deleted test files.")
-
-    os.remove(weirdPath)
     
-    print("Splitting chapters:")
+    if os.path.exists(weirdPath):
+        os.remove(weirdPath)
 
     splitChapter(fullPath, 20)
-
-    # tts = TTSProcessor()
 
     for root, _, files in os.walk(savePath):
         print("Processing on each file:")
@@ -33,8 +31,6 @@ def remakeTestFile():
             filePath = os.path.join(root, filename)
             try:
                 deleteLine(filePath)
-
-                # generateTTS(filePath, tts)
 
                 addHeading(filePath)
                 addCodeBlock(filePath)
